--- openai/pyproject.toml
@@ -1,23 +1,0 @@
-[tool.poetry]
-name = "integrations-python-openai"
-version = "0.0.1"
-description = "Python Restack AI OpenAI integration"
-authors = [
-    "Restack Team <service@restack.io>"
-]
-readme = "README.md"
-packages = [
-    { include = "integrations_python_openai" }
-]
-include = ["src/py.typed"]
-
-[tool.poetry.dependencies]
-python = "^3.8"
-openai = "^1.52.0"
-pydantic = "^2.9.2"
-restack-ai = "^0.0.11"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
