+++ restackio-integrations-openai/pyproject.toml
@@ -1,0 +1,21 @@
+[tool.poetry]
+name = "restackio.integrations.openai"
+version = "0.0.1"
+description = "OpenAI integration for Restack AI"
+authors = [
+    "Restack Team <service@restack.io>"
+]
+packages = [{ include = "restackio", from = "src" }]
+include = ["src/restackio/integrations/openai/py.typed"]
+
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+restackio.ai = "^0.0.11"
+pydantic = "^2.9.2"
+openai = "^1.52.0"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
