+++ restackio-integrations-openai/src/restackio/integrations/openai/__init__.py
@@ -1,4 +1,7 @@
-from .functions.chat.completions_base import OpenAIChatInput, OpenAIChatOutput, openai_chat_completion_base
+# src/restackio/integrations/openai/__init__.py
+__path__ = __import__('pkgutil').extend_path(__path__, __name__)
+
+from .functions.chat.completions_base import openai_chat_completion_base, OpenAIChatInput, OpenAIChatOutput
 from .utils.client import openai_client
 from .utils.cost import TokensCount, Price, openai_cost
 from .service import OpenAIServiceOptions, OpenAIServiceInput, openai_service
