+++ restackio-integrations-openai/src/restackio/integrations/openai/functions/chat/completions_base.py
