+++ restackio-integrations-openai/src/restackio/integrations/openai/service.py
