+++ restackio-integrations-openai/src/restackio/integrations/openai/utils/client.py
