+++ restackio-integrations-toolhouse/src/restackio/integrations/toolhouse/__init__.py
@@ -1,0 +1,2 @@
+# src/restackio/integrations/toolhouse/__init__.py
+__path__ = __import__('pkgutil').extend_path(__path__, __name__)
