+++ restackio-integrations-toolhouse/src/restackio/integrations/toolhouse/functions/send_email.py
@@ -1,22 +1,24 @@
 from pydantic import BaseModel
 from restack_ai.function import function
 from openai import OpenAI
-from restack_integrations_toolhouse.utils.toolhouse_client import toolhouse_client
+from ..utils.toolhouse_client import toolhouse_client
 
-class WebsiteInput(BaseModel):
-    url: str
+class MailInput(BaseModel):
+    subject: str
+    body: str
+    to: str
     openai_api_key: str | None = None
     toolhouse_api_key: str | None = None
 
 
-@function.defn(name="summarize_website")
-async def summarize_website(input: WebsiteInput) -> str:
+@function.defn(name="mail_website_summary")
+async def mail_website_summary(input: MailInput) -> str:
     client = OpenAI(api_key=input.openai_api_key)
     th = toolhouse_client(api_key=input.toolhouse_api_key, provider="openai")
 
     messages = [
-        {"role": "user", "content": f"Get the contents of {input.url} and summarize its key value propositions in three bullet points."},
-    ]    
+        {"role": "user", "content": f"Send email to {input.to} with the following subject: {input.subject} and body: {input.body}"},
+    ]   
 
     response = client.chat.completions.create(
         model="gpt-4o",
@@ -28,10 +30,10 @@ async def summarize_website(input: WebsiteInput) -> str:
 
     print(messages)
 
-    response = client.chat.completions.create(
+    client.chat.completions.create(
         model="gpt-4o",
         messages=messages,
         tools=th.get_tools()
     )
 
-    return response.choices[0].message.content
+    return "Email sent successfully"
