--- toolhouse/pyproject.toml
@@ -1,20 +1,0 @@
-[tool.poetry]
-name = "restack_integrations_toolhouse"
-version = "0.1.0"
-description = "Python Restack AI Toolhouse integration"
-authors = [
-    "Restack Team <service@restack.io>"
-]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = ">=3.10,<3.13"
-toolhouse = "^1.3.0"
-restack-ai = "^0.0.11"
-openai = "^1.52.1"
-pydantic = "^2.9.2"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
